Complete Refactoring: Using new build system
* Now using ImgLib 2.2.0 base (still open problems)
* No need to commit any libraries anymore. We generate bundles
* Testflows adapted but still fail: Waiting for bug-fixes in ImgLib2

+++ org.knime.knip.core/src/org/knime/knip/core/io/externalization/externalizers/ImgExt0.java
@@ -389,7 +389,7 @@ public class ImgExt0 implements Externalizer<Img> {
                 final Cursor<Unsigned12BitType> unsigned12BitTypeCursor = (Cursor<Unsigned12BitType>)cur;
                 while (cur.hasNext()) {
                     cur.fwd();
-                    out.writeShort(unsigned12BitTypeCursor.get().get());
+                    out.writeLong(unsigned12BitTypeCursor.get().get());
                 }
                 break;
             case UNSIGNEDBYTETYPE:
